Some haphazard stabs at GANs,
So I was tasked with making a model using GANs. The model would ideally be able to generate good synthetic datasets of EEG data. I did a little work generating artificial photos, this showed me that hyper parameter optimisation is a bitch and its really tough to keep loss in the generator high as once there is a discrepancy between it and the discriminator the loss collapses. 

I had a plan to use Fourier transforms to convert sections of the VEP dataset signals into functions and then create a model that could build chains of those functions, I made a model that extracts VEP signals from the dataset. 
In order to do this I used ICA to try clean the dataset of some of its inherent noises. Removing particularly distorted frequencies lead to a clearer channel. 


Now I plan to make a set of functions that compare models trained to extract VEPs from data with different frequency combinations removed. 
I plan on turning the average VEP response into a paramatizable function. 
I plan to make a diffusion model which generates VEP signal data, maybe a GAN with the generator as a diffusion model. 
I want to work on genetic elvolution optimisation of my diffusion model. (honestly someone was telling me about it and I thought it was cool)